feat/create-wallet: P0/P1/P2 bugs fixed

+++ src/main/java/org/brunagabriel/service/WalletService.java
@@ -4,8 +4,10 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.NotFoundException;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
 
 import org.brunagabriel.dto.WalletRequest;
 import org.brunagabriel.dto.CreateWalletResponse;
@@ -13,22 +15,20 @@ import org.brunagabriel.dto.WalletResponse;
 import org.brunagabriel.entity.WalletEntity;
 import org.brunagabriel.repository.WalletRepository;
 
-@ApplicationScoped //deixa a classe pronta para que seja usada facilmente - cria a classe uma unica vez -singleton
+@ApplicationScoped
 public class WalletService {
 
-    @Inject
-    private WalletRepository repository;
+    private final WalletRepository repository;
 
-//    @Transactional //este metodo tem de rodar dentro de uma transação
-//    public WalletResponse save (WalletRequest request){
-//
-//        WalletEntity wallet= WalletEntity.toWallet(request);
-//        repository.persist(wallet);
-//        return WalletResponse.fromWallet(wallet);
-//    }
+    public WalletService(WalletRepository repository){
+        this.repository = repository;
+    }
 
     @Transactional
     public CreateWalletResponse save (WalletRequest request){
+        if (repository.findByEmail(request.email())){
+            throw new WebApplicationException("Email already exists", Response.Status.CONFLICT);
+        }
         WalletEntity wallet=new WalletEntity();
         wallet.setEmail(request.email());
         wallet.setBalance(request.amount());
@@ -40,6 +40,10 @@ public class WalletService {
     @Transactional
     public WalletResponse searchById(Long id) {
         WalletEntity wallet = repository.findById(id);
+
+        if(wallet==null){
+            throw new NotFoundException("Wallet not found");
+        }
         return WalletResponse.fromWallet(wallet);
     }
 
@@ -54,6 +58,10 @@ public class WalletService {
     @Transactional
     public WalletResponse update(Long id, WalletRequest request) {
         WalletEntity wallet = repository.findById(id);
+
+        if (wallet==null){
+            throw new NotFoundException("Wallet not found");
+        }
         wallet.setEmail(request.email());
         return WalletResponse.fromWallet(wallet);
     }
@@ -61,6 +69,9 @@ public class WalletService {
     @Transactional
     public void delete(Long id) {
         WalletEntity wallet = repository.findById(id);
+        if (wallet==null){
+            throw new NotFoundException("Wallet not found");
+        }
         repository.delete(wallet);
     }
 }
